architekt-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
bcrypt
PyJWT
